% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochita.R
\name{mochita}
\alias{mochita}
\title{Test an HTTP server}
\usage{
mochita(app)
}
\arguments{
\item{app}{App definition, see \link[httpuv:startServer]{httpuv::startServer}.}
}
\value{
A \code{Mochita} R6 class object.
}
\description{
\code{mochita()} builds the specification for testing an \code{httpuv} app.
It returns a \code{Mochita} object whose methods can be chained together
to build up a request, describe the expectations it
should meet, and finally send it with \verb{$perform()}.
}
\section{Mochita Methods}{


A \code{Mochita} is an R6 Class with the following methods:
\itemize{
\item \verb{$get(path)}, \verb{$head(path)}, \verb{$post(path)}, \verb{$put(path)},
\verb{$delete(path)}, \verb{$connect(path)}, \verb{$options(path)}, \verb{$trace(path)},
\verb{$patch(path)} Set the HTTP Method and path of the request.
\item \verb{$set(name,value)} Set a request header.
\item \verb{$expect()} Add an expectation to check against the response,
once performed. Accepts a status code (optionally with a body to match),
a header name/value pair, a body to match, or a custom function taking
the response.
\item \verb{$perform()} Send the request and run the recorded expectations against
the response.
}
}

\examples{
\dontrun{
mochita(app)$get("/")$
  set("Accept", "text/plain")$
  expect(200)$
  expect("Content-Type", "text/plain")$
  perform()
}

}
