% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochita-package.R
\docType{package}
\name{mochita-package}
\alias{mochita-package}
\title{mochita: Test R Web Applications}
\description{
Write so-called 'Integration Tests' for your R web applications by declaring an HTTP request and the expectations its response should meet.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JulioCollazos64/mochita}
  \item Report bugs at \url{https://github.com/JulioCollazos64/mochita/issues}
}

}
\author{
\strong{Maintainer}: Julio Collazos \email{amarullazo626@gmail.com}

Authors:
\itemize{
  \item Julio Collazos \email{amarullazo626@gmail.com}
}

Other contributors:
\itemize{
  \item  supertest contributors (authors listed in <https://github.com/forwardemail/supertest/graphs/contributors>) [contributor, copyright holder]
}

}
\keyword{internal}
